recoded validation of forms
