+++ index.html
@@ -17,9 +17,9 @@
                     
                 </div>
             </div>
-            <script type="module" src="script.js"></script>
-            <script type="module" src="validationForm.js"></script>
             <script type="module" src="route.js"></script>
+            <!--<script type="module" src="validationForm.js"></script>-->
+            <script type="module" src="script.js"></script>
         </div>
         
     </body>
