+++ styles.css
@@ -25,7 +25,8 @@
     min-height: 400px;
     width: 420px;
     background-color: rgba(248, 239, 239, 0.75);
-    margin-top: 15%;
+    margin-top: 10%;
+    margin-bottom: 10%;
     padding-bottom: 5%;
     border-radius: 30px;
     outline: none;
@@ -55,6 +56,7 @@
     background-color: transparent;
     margin-top: 5%;
     width: 80%;
+    height: 100%;
     
 }
